This shaves another 10 to 20ms of the total runtime of 300ms in my testcase. Again this is a small improvement but they add up :)