@cocreature 2016-10-05
> You've probably remembered to check functional tests too ? ("make test")

I think so, I’ll run them again just to be sure. Could we enable them on Travis? Tests that have to be run manually have the tendency to get broken easily.

> What determines which functions get the INLINABLE pragma ? What would be the policy going forward ?

I mostly looked at the [flamegraph](https://github.com/fpco/ghc-prof-flamegraph.git) of a profile and checked which functions seem worth inlining and put INLINABLE pragmas on those. There are three criterions that I generally use

1. Is this function called only once? In this case inlining doesn’t increase code size and often enables further optimizations.

2. Is this function polymorphic? Inlinable exposes the unfolding so functions can be specialized.

3. Is this function small and called from a lot of different modules?

None of these are hard criterions and in the end I just guess and see what seems to make an effect. There are probably some unnecessary inlinables (as in the performance doesn’t improve significantly. They all seemed to inline after adding the pragma) and I’m sure there are other places where we could gain something. I don’t think a policy is useful, the decision has to be made on a case by case basis.

> Could you comment on SO answers like http://stackoverflow.com/questions/21356897/do-i-need-to-bother-with-inline-inlinable-pragmas-for-small-exported-functions and http://stackoverflow.com/questions/15440804/disadvantages-of-using-inlinable-pragma ? Are things like build time, build memory usage, executable size affected much ?

I don’t feel qualified enough to answer these especially given that people actually working on GHC have already answered. I haven’t done any measures on the build and the executable size because tbh I’m more than willing to accept longer build times for better runtime performance. If you are looking for better build time performance there are bigger fish to fry, e.g. isolating the TH in `hledger` to avoid building almost every module twice.