@cocreature 2016-10-05
Here are the benchmarks:

#### before
```
benchmarking read bench/10000x1000x10.journal
time                 2.001 s    (1.982 s .. 2.023 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.995 s    (1.991 s .. 1.998 s)
std dev              4.042 ms   (0.0 s .. 4.658 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking print
time                 1.651 s    (1.566 s .. 1.718 s)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.637 s    (1.623 s .. 1.647 s)
std dev              16.23 ms   (0.0 s .. 18.25 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 2.088 s    (1.993 s .. 2.157 s)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 2.088 s    (2.077 s .. 2.097 s)
std dev              15.57 ms   (0.0 s .. 16.29 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking balance
time                 211.3 ms   (207.2 ms .. 213.2 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 208.0 ms   (206.5 ms .. 209.3 ms)
std dev              1.708 ms   (981.7 μs .. 2.203 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking stats
time                 194.8 ms   (186.3 ms .. 201.7 ms)
                     0.998 R²   (0.993 R² .. 1.000 R²)
mean                 195.8 ms   (191.0 ms .. 198.6 ms)
std dev              4.466 ms   (2.133 ms .. 6.387 ms)
variance introduced by outliers: 14% (moderately inflated)
```

#### after
```
Benchmarking hledger in /home/moritz/code/haskell/hledger/hledger with criterion
benchmarking read bench/10000x1000x10.journal
time                 1.842 s    (1.824 s .. 1.862 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.829 s    (1.820 s .. 1.833 s)
std dev              7.411 ms   (0.0 s .. 8.004 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking print
time                 1.605 s    (1.541 s .. 1.688 s)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 1.607 s    (1.594 s .. 1.618 s)
std dev              16.00 ms   (0.0 s .. 18.12 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking register
time                 2.181 s    (2.175 s .. 2.191 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.148 s    (2.137 s .. 2.154 s)
std dev              10.03 ms   (0.0 s .. 11.33 ms)
variance introduced by outliers: 19% (moderately inflated)

benchmarking balance
time                 212.9 ms   (207.0 ms .. 217.8 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 214.1 ms   (212.3 ms .. 216.3 ms)
std dev              2.491 ms   (1.684 ms .. 3.335 ms)
variance introduced by outliers: 14% (moderately inflated)

benchmarking stats
time                 190.2 ms   (162.4 ms .. 199.7 ms)
                     0.991 R²   (0.971 R² .. 1.000 R²)
mean                 204.2 ms   (195.0 ms .. 229.3 ms)
std dev              18.65 ms   (1.565 ms .. 24.39 ms)
variance introduced by outliers: 16% (moderately inflated)

Benchmark bench: FINISH
```

I ran the tests a few times and I’ve always seen a similar improvement on parsing time while the other measurements seem meaningless (sometimes one won sometimes the other).