@simonmichael 2016-10-12
Thanks for your comments and results, this is valuable exploration. I feel cautious about merging clever GHC management like this because I think we can quickly get into diminishing returns, with improvements that are fragile, requiring ongoing testing. 

I am doing some testing also. So far I see these INLINABLE pragmas causing no change in executable size, no change in runtime memory usage, about an 8% compilation slowdown, and a possible slight speedup in parsing.

I have to say "possible" because I'm having trouble measuring definitively. I'm looking at ways to make this easier, which we need generally. (I'm getting good value from the quick, low detail, comparative reports of tools/simplebench.hs, which I think I'll improve and package separately as quickbench.)